Refactor sleep mixin to be self contained Hooking into the core's mechanicsism and ensuring initialisation of further mixins, cleaning up if last.
